+++ packages/wujie-core/src/common.ts
@@ -111,6 +111,7 @@ export const documentProxyProperties = {
   // 需要从主应用document中获取的方法
   documentMethods: [
     "execCommand",
+    "caretPositionFromPoint",
     "createRange",
     "exitFullscreen",
     "exitPictureInPicture",
@@ -166,6 +167,8 @@ export const appWindowAddEventListenerEvents = [
   "beforeunload",
   "unload",
   "message",
+  "error",
+  "unhandledrejection",
 ];
 
 // 子应用window.onXXX需要挂载到iframe沙箱上的事件