+++ packages/wujie-react/index.js
@@ -29,13 +29,18 @@ export default class WujieReact extends React.PureComponent {
       console.log(error);
     }
   }
+  
+  componentDidMount () {
+    this.startApp();
+  }
 
-  execStartApp = () => {
-    this.startAppQueue = this.startAppQueue.then(this.startApp);
-  };
+  componentDidUpdate(prevProps) {
+    if (this.props.name !== prevProps.name || this.props.url !== prevProps.url) {
+      this.startApp();
+    }
+  }
 
   render() {
-    this.execStartApp();
     const { width, height } = this.props;
     const { myRef: ref } = this.state;
     return <div style={{ width, height }} ref={ref} />;